feat: migrate main files
