+++ package.json
@@ -10,6 +10,7 @@
   "main": "src/server.js",
   "scripts": {
     "build": "tsc",
+    "start": "node --loader ts-node/esm --es-module-specifier-resolution=node src/server.ts",
     "lint": "eslint --ext .js,.ts --fix src/",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
     "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace"
