+++ src/server.ts
@@ -1,5 +1,5 @@
-import app from './app.js';
-import config from './common/config.js';
+import app from './app';
+import config from './common/config';
 
 const { PORT, NODE_ENV } = config;
